refactor: update imports after common module split
