+++ src/pynamicalsys/discrete_time/gali.py
@@ -20,7 +20,7 @@ from numba import njit
 from numpy.typing import NDArray
 from typing import Tuple
 from pynamicalsys.common.types import int_t, numeric_t, map_t, jacobian_t
-from pynamicalsys.common.utils import qr
+from pynamicalsys.common.linalg import qr
 
 
 @njit
