+++ src/pynamicalsys/hamiltonian_systems/gali.py
@@ -20,7 +20,7 @@ import numpy as np
 from numba import njit
 
 from pynamicalsys.common.types import grad_t, hess_t, symplectic_tangent_step_t
-from pynamicalsys.common.utils import qr
+from pynamicalsys.common.linalg import qr
 
 
 @njit
